Automatically resize the branding box (and logo) with the sidebar
This makes it possible to use any branding and any sidebar size
transparently, making it easier to fluidify the layout later.

The current downside is that we fix the box as a square, which may
add some bottom margin to the image depending on its aspect ratio.
To fix this we would need to read the aspect ratio of the image and
it in '%' as the padding-top style attribute, but I think it is not
worth it.