CSS: Add a new "FlexRonRon" stylesheet prototype.
This new stylesheet is designed to be exactly the same as RonRonnement,
but with fonts automatically growing in size depending on the width of
the viewport.

To achieve this, variables in RonRonnement have been made overrideable.

Currently it uses the Open Sans font from Google Fonts because the
default sans-serif font looks very bad at bigger sizes.